move features contexts to subdir
