+++ packages/core/react/src/features/ConnectProvider.tsx
@@ -3,7 +3,7 @@ import type { IdentifierRecord } from '@wallet-standard/standard';
 import type { FC, ReactNode } from 'react';
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { ConnectContext } from './useConnect.js';
-import { useWallet } from './useWallet.js';
+import { useWallet } from '../useWallet.js';
 
 /** TODO: docs */
 export interface ConnectProviderProps {
