+++ packages/core/react/src/features/index.ts
@@ -1,0 +1,2 @@
+export * from './ConnectProvider.js';
+export * from './useConnect.js';
