+++ packages/core/react/src/index.ts
@@ -1,5 +1,5 @@
-export * from './ConnectProvider.js';
-export * from './useConnect.js';
+export * from './features/index.js';
+
 export * from './useWallet.js';
 export * from './useWallets.js';
 export * from './WalletProvider.js';
